fixed tokens

+++ models/contact.go
@@ -16,7 +16,7 @@ import (
 // Contact gets the Name, phone and UserID of the contact
 type Contact struct {
 	gorm.Model
-	UserID  uint     `json:"userId"` //The user that this contact belongs to
+	UserID  uint64   `json:"userId"` //The user that this contact belongs to
 	Name    FullName `gorm:"foreignkey:ContactID"`
 	Address Address  `gorm:"foreignkey:ContactID"` // Mailing address
 	//Properties []Address `json:"propertyAddresses" gorm:"foreignKey:ID"` // The properties owned by the owner
